+++ Changelog.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
  - ObjectBuilder directive.
 
+### Added
+ - 'useLocaleSettings' option to Day and DateTime components.
+ - Label setting for HTML element component
+
 ## 2.26.0
 ### Changed
  - Upgrade ng-formio to 2.26.0